Add Idealised Value Metric (#4414)
This Pr adds a new metric (idealised_scheduled_value) which represents
the "ideal" value of jobs that could be scheduled each scheduling cycle.

We calculate this by considering that all (pool) resources are present
on one giang node. This frees us from pesky constraints such as "we need
to binpack efficiently" and allows us to quickly generate an upper bound
for the maximum amount of value that could be scheduled.

---------

Signed-off-by: Chris Martin <chris@cmartinit.co.uk>
Co-authored-by: Christopher Martin <Chris.Martin@gresearch.co.uk>