Lookout UI: add Sentry error monitoring support and load config from JS file (#4426)
This adds options to the Lookout UI configuration to use error
monitoring using Sentry, with the possibility to easily further extend
this to other error monitoring providers.

The error handlers are passed to React's `createRoot()` function, which
ensures all errors which are thrown anywhere inside the React
application are processed and sent to the configured error monitoring
providers, if any.

For sending such errors to Sentry, the following configuration would
need to be added for Lookout:
```
uiConfig:
  errorMonitoring:
    sentry:
      dsn: "https://key@sentry.url/project"
      environment: "your-env"
  ```

In the Lookout UI, we currently load UI config, ultimately defined by the party deploying and operating the Armada instance, via an HTTP request to the GET /config endpoint,

This commit removes that endpoint, instead loading this configuration via a JavaScript file generated by the server, which is imported directly by the index.html. The JavaScript file writes the configuration into a property of the browser's `window` object.

This is advantageous because:
- we don't perform a `fetch()` call at runtime in the React app and so can remove the code and maintenance this requires
- we can use configuration values when initiating the React app