Fix unhandled request warning in Lookout UI tests (#4425)
When running the tests in `Sidebar.test.tsx`, we get the following error
because the fetch requests to retrieve the job run error and job run
debug message are not being mocked:
```
[MSW] Error: intercepted a request without a matching request handler:

  • POST /api/v1/jobRunError

  • Request body: {"runId":"1234-5678"}

If you still wish to intercept this unhandled request, please create a request handler for it.
```

This PR removes that warning by creating the request handlers.