Lookout UI: enable grouping by namespace (#4445)
Adds support for grouping by namespace, so long as queue is already
being grouped-by.

This also makes filtering on the queue column a prerequisite to
filtering by the namespace column, and enforces all filtering
prerequisites in the group-by selector.

Add an index on job (queue, namespace) to enhance the query performance
for filtering and grouping by namespace.