Interning resource requirement strings (#450) (#4468)
<!--  Thanks for sending a pull request!  Here are some tips for you:

#### What type of PR is this?

#### What this PR does / why we need it:

#### Which issue(s) this PR fixes:
<!--
*Automatically closes linked issue when PR is merged.
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
_If PR is about `failing-tests or flakes`, please post the related
issues/tests in a comment and do not use `Fixes`_*
-->


This PR interns more of the strings that we use when inserting a job
into the jobDB, such as priority class and pools, as well as those in
the k8s resource requirements.

Co-authored-by: Mustafa Ilyas <Mustafa.Ilyas@gresearch.co.uk>