chore: update golangci-lint (#4410)
#### What type of PR is this?

Chore PR which updates golangci-lint to newer version (v2).

Updates minimal version of golangci-lint to v2.5.0 in mage.
Updates CI to use v8 version of the golangci-lint action and v2.5.0 of
golangci-lint.
Updates golangci-lint config to v2 format using `golangci-lint migrate`.

#### What this PR does / why we need it:

golangci-lint has moved to a new major version and the config format is
not compatible.

Signed-off-by: Dejan Zele Pejchev <pejcev.dejan@gmail.com>