Allow pods to be retried when failed containers match regex (#4483)
This adds functionality to the executor to allow Armada to be configured
such that pods that have failed and have entered "user" code, i.e any
initContainers or Containers in non-pending state, can be retried if
they match configured regex.

This regex is for the container name and the container message. For any
retrying of this type to occur, both the container name regex and the
message regex must match. Given the nature of regex, not setting the
regex value results in that regex matching all string values.
For example, this configuration would retry any container where 'test'
is in the termination message, regardless of the container name.
```
failedPodChecks:
  containerStatuses:
      messageRegexp: "test"
```
whereas this configuration would match any container, regardless of what
the termination message is or if there is a termination message value at
all. (The message string can be empty.)
```
failedPodChecks:
  containerStatuses:
    - containerNameRegexp: ".*"
```

---------

Co-authored-by: Eleanor Pratt <Eleanor.Pratt@gresearch.co.uk>