[Lookout] allow configuration of job links in the sidebar (#4512)
Allows configuration for link chips to appear at the top of the job
table sidebar, based on values from the job spec, such as annotation
values. This is useful for better integrating with external tooling,
such as CI/CD. workflow orchestration, MLOps and data processing.

Links will not be rendered for jobs which do not have values present,
which are required to generate a link according to the configuration.

For example, we can add this to the Lookout configuration:
```yaml
uiConfig:
  jobLinks:
    - label: Apache Airflow Job
      colour: "#00c7d4"
      linkTemplate: 'https://your-airflow.internal/dags/{{ annotations["armadaproject.io/apache-airflow-dag-id"] }}/grid?dag_run_id={{ annotations["armadaproject.io/apache-airflow-dag-run-id"] }}&task_id={{ annotations["armadaproject.io/apache-airflow-task-id"] }}'
    - label: Jenkins build
      colour: "#335061"
      linkTemplate: 'https://your-jenkins.internal/job/{{ annotations["armadaproject.io/jenkins-job-name"] }}/{{ annotations["armadaproject.io/jenkins-build-number"] }}'
    - label: MLflow run
      colour: "#3578e5"
      linkTemplate: 'https://your-mlflow.internal/#/experiments/{{ annotations["armadaproject.io/mlflow-experiment-id"] }}/runs/{{ annotations["armadaproject.io/mlflow-run-id"] }}'
```

It would cause Lookout to render three link chips in the sidebar for a
job with these annotations:
```yaml
jobs:
  - annotations:
      armadaproject.io/apache-airflow-dag-id: my_dag_id
      armadaproject.io/apache-airflow-dag-run-id: run_20250101
      armadaproject.io/apache-airflow-task-id: "my_task"
      armadaproject.io/jenkins-job-name: "my_job"
      armadaproject.io/jenkins-build-number: "1234"
      armadaproject.io/mlflow-experiment-id: "5"
      armadaproject.io/mlflow-run-id: abcdabcd-0000-0000-0000-0123456789ab
```

<img width="868" height="892" alt="image"
src="https://github.com/user-attachments/assets/8aa0b1af-0a2e-407f-9c87-77936fe45bfb"
/>