Publish test results for fork PRs (#4336)
#### What type of PR is this?
Purely CI

#### What this PR does / why we need it:
Test results of pull requests from fork repositories are currently not
published. The current setup makes them fail:

    Resource not accessible by integration

https://github.com/armadaproject/armada/actions/runs/14449584216/job/40518701460#step:5:2282

This is because the CI is executed in the context of the fork
repository, where Github Actions does not have permissions to write to
the upstream pull request (required by the publish action).

#### Which issue(s) this PR fixes:
This applies the pattern documented at
https://github.com/mikepenz/action-junit-report#pr-run-permissions.

---------

Signed-off-by: Enrico Minack <github@enrico.minack.dev>
Signed-off-by: Nikola Jokic <jokicnikola07@gmail.com>